requirements.txt
mido==1.2.9
click==7.1.2
RPi.GPIO==0.7.0
